DEV-53 Point: 1, 2, 3, 6, 9
