+++ .env.example
@@ -1,7 +1,7 @@
-VITE_FIREBASE_API_KEY
-VITE_FIREBASE_AUTH_DOMAIN
-VITE_FIREBASE_DATABASE_URL
-VITE_FIREBASE_PROJECT_ID
-VITE_FIREBASE_STORAGE_BUCKET
-VITE_FIREBASE_MESSAGING_SENDER_ID
-VITE_FIREBASE_APP_ID
+VITE_FIREBASE_API_KEY=AIzaSyBjycuYIdU9dHr3K321x_KbPBH56dZOC_s
+VITE_FIREBASE_AUTH_DOMAIN=skyfitnespro.firebaseapp.com
+VITE_FIREBASE_DATABASE_URL=https://skyfitnespro-default-rtdb.europe-west1.firebasedatabase.app
+VITE_FIREBASE_PROJECT_ID=skyfitnespro
+VITE_FIREBASE_STORAGE_BUCKET=skyfitnespro.appspot.com
+VITE_FIREBASE_MESSAGING_SENDER_ID=1032692405656
+VITE_FIREBASE_APP_ID=1:1032692405656:web:745f61781f272d8ffa47ed
