+++ index.html
@@ -7,7 +7,7 @@
     <link href="https://fonts.googleapis.com/css2?family=Roboto:wght@400;500;700&display=swap" rel="stylesheet">
     <title>Skypro Fitnes</title>
   </head>
-  <body>
+  <body style="background-color: #FAFAFA">
     <div id="root"></div>
     <script type="module" src="/src/main.tsx"></script>
   </body>
