+++ src/components/Cards/MyCourseCard.tsx
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from 'react';
 import { useNavigate } from 'react-router-dom';
 import Button from "../Button";
 import Level from "./Level";
-import WorkoutListModal from '../WorkoutListModal';
+import WorkoutListModal from '../Modals/WorkoutListModal';
 import { Course, Workout } from '../../types/interfaces';
 import { database } from '../../config/firebase';
 import { ref, get, set } from "firebase/database";
@@ -89,18 +89,17 @@ const MyCourseCard: React.FC<MyCourseCardProps> = ({ course, onCourseRemoved }) 
   };
 
   return (
-    <div className="w-[360px] pb-[15px] rounded-[30px] shadow-[0px_4px_67px_-12px_rgba(0,0,0,0.13)]">
+    <div className="w-full pb-[15px] rounded-[30px] shadow-[0px_4px_67px_-12px_rgba(0,0,0,0.13)]">
       <div className="relative" onClick={() => navigate(`/course/${course._id}`)}>
         <img 
           src={`/images/${course.nameEN}.png`}
           alt={course.nameRU} 
-          className="w-[360px] h-[325px] rounded-[30px] object-cover"
+          className="w-full h-[325px] rounded-[30px] object-cover"
           loading="lazy"
         />
         <button className="absolute top-5 right-5 w-[32px] h-[32px]" onClick={handleRemoveCourse}>
           <svg width="32" height="32" viewBox="0 0 32 32" fill="none" xmlns="http://www.w3.org/2000/svg">
-            <circle cx="16" cy="16" r="14" fill="white" />
-            <path d="M10 16h12" stroke="black" strokeWidth="2" />
+            <path fill-rule="evenodd" clip-rule="evenodd" d="M16 29.3333C23.3638 29.3333 29.3333 23.3638 29.3333 16C29.3333 8.63616 23.3638 2.66663 16 2.66663C8.63619 2.66663 2.66666 8.63616 2.66666 16C2.66666 23.3638 8.63619 29.3333 16 29.3333ZM14.6667 14.6666V9.33329H17.3333V14.6666H22.6667V17.3333H17.3333V22.6666H14.6667V17.3333H9.33332V14.6666H14.6667Z" fill="white"/>
           </svg>
         </button>
       </div>
@@ -136,7 +135,7 @@ const MyCourseCard: React.FC<MyCourseCardProps> = ({ course, onCourseRemoved }) 
         </div>
         <Button 
           variant='primary' 
-          className='w-[300px] h-[52px] mt-[40px]' 
+          className='w-full h-[52px] mt-[40px]' 
           onClick={() => handleOpenModal(progress === 0 ? 'start' : progress === 100 ? 'restart' : 'continue')}
         >
           {getButtonText()}
