+++ src/pages/ProfilePage/Profile.tsx
@@ -1,5 +1,5 @@
 import { useState } from 'react'
-import Modal from '../../components/Modal'
+import Modal from '../../components/Modals/Modal'
 import { useAuth } from '../../hooks/useAuth'
 
 function Profile() {
